finished vite migration and restructured scss
