+++ frontend/src/Components/Article/Article.scss
@@ -1,3 +1,4 @@
+
 @import "../../node_modules/bootstrap/scss/bootstrap-grid.scss";
 @import "../../node_modules/bootstrap/scss/grid";
 @import "../../node_modules/bootstrap/scss/variables";
