+++ frontend/src/Components/Article/ArticleComponents.scss
@@ -1,3 +1,5 @@
+@import '../../scss/variables';
+
 //SectionHeader, ArticleHeaderDesc marker styling
 .section-header-marker,
 .article-desc-marker {
