+++ frontend/src/Components/Article/Section/SectionHeader.jsx
