+++ frontend/src/Components/LinkedDescriptionBox.jsx
