+++ frontend/src/Components/Signin/Signin.scss
@@ -1,4 +1,9 @@
-$signin_theme:#B4DBF1;
+@import "../../node_modules/bootstrap/scss/bootstrap-grid.scss";
+@import "../../node_modules/bootstrap/scss/grid";
+@import "../../node_modules/bootstrap/scss/variables";
+@import "../../node_modules//bootstrap/scss/mixins";
+
+$signin_theme: #b4dbf1;
 
 .bg-signin-card {
   background-color: $signin_theme;
@@ -25,4 +30,3 @@ $signin_theme:#B4DBF1;
     width: 350px;
   }
 }
-
