+++ frontend/src/Components/Signup/Signup.scss
@@ -1,3 +1,8 @@
+@import "../../node_modules/bootstrap/scss/bootstrap-grid.scss";
+@import "../../node_modules/bootstrap/scss/grid";
+@import "../../node_modules/bootstrap/scss/variables";
+@import "../../node_modules//bootstrap/scss/mixins";
+
 .w-signup {
   @include media-breakpoint-down(md) {
     max-width: 350px;
