+++ frontend/src/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import "./css/custom.css";
+import "./scss/custom.scss"
 import { BrowserRouter } from "react-router-dom";
 import reportWebVitals from "./reportWebVitals";
