+++ frontend/src/pages/BasePage.jsx
