+++ frontend/src/pages/SignPages/SignPages.scss
@@ -1,0 +1,5 @@
+@import '../../scss/variables';
+.sign-page-background{
+    background: linear-gradient(180deg, rgba(230, 231, 230, 0.40) 47.57%, rgba(221, 206, 206, 0.00) 100%), $primary;
+    min-height:95vh;
+}
