--- frontend/src/scss/SignPages.scss
@@ -1,4 +1,0 @@
-.sign-page-background{
-    background: linear-gradient(180deg, rgba(230, 231, 230, 0.40) 47.57%, rgba(221, 206, 206, 0.00) 100%), #F65757;
-    min-height:95vh;
-}
