+++ frontend/src/scss/_variables.scss
@@ -1,0 +1,2 @@
+$primary: #f65757;
+$secondary: #228b22;
